pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gdou-dianfei"
version = "0.1.0"
description = "广东海洋大学宿舍电费查询客户端"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "彼岸", email = "farshore07@qq.com"}
]
keywords = ["gdou", "electricity", "balance"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "pycryptodome>=3.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
include = ["gdou_dianfei*"]
